test: enhance useAppShellController tests with activity tracking and state restoration
- Added tests for theme changes and activity recording in settings.
- Verified activity entries in the Activity drawer model.
- Ensured activity issue indicators clear only after opening Activity.
- Recorded Firestore query activity and handled invalid saved workspace states.
- Implemented tests for closing tabs with running queries and maintaining state.
- Added tests for document path queries, auth filtering, and loading users.
- Verified routing of document saves, creates, updates, and deletes through Firestore repository.
- Enhanced test structure with utility functions for better readability and maintainability.
